platform/x86/intel/ifs: Track and report scan results on a per-cpu basis
IFS requires invoking the scan simultaneously as indicated in the
INTEGRITY_MSR (bits 20:16), when bit 21 is set, that indicates for all
those thread groups need to trigger the test at the same time.

When the tests are run simultaneously, record the test results in a per-cpu
structure. This reduces the need to repeat the test redundantly for all the
shared cores.

There are some benefits:

- Reduced power spent, by avoiding the scan multiple times.
- Reduces the time to complete a scan by 50% or more.
- These tests could take upto 200ms. Prevents taking these cores out of
circulation just to repeat tests that have already been tested.

Signed-off-by: Jithu Joseph <jithu.joseph@intel.com>