KVM: x86/pmu: Fixed the availability issue of guest Perfmon V6 MSRs
These MSR aliases are supposed to be exposed to the guest only if KVM
exposes Architectural Perfmon version 6+ to the guest.  e.g. if the
host supports V6, but the guest Perfmon version is set to 2 because
mediated vPMU is disabled for the guest, KVM should intercept these
V6 MSRs and return false from intel_is_valid_msr().

Fixes: 88c31023c522 ("KVM: x86/pmu: Add Perfmon MSR aliasing support for vPMU")
Fixes: 4d1a68e0d4e0 ("KVM: x86/pmu: Pass-through v6 new range MSRs")
Signed-off-by: Zide Chen <zide.chen@intel.com>