KVM: x86/pmu: Refactor some MSR interception code
intel_pmu_update_msr_intercepts() could be called multiple times with
different PMU configurations, and the interception setting for every MSR
needs to be refreshed.

Even for MSRs that are not impacted by userspace re-configuration,
removing the conditions to invoke the vmx_set_intercept_for_msr()
makes the code simple.

Signed-off-by: Zide Chen <zide.chen@intel.com>