KVM: x86/kvm: Don't expose CPUID.23H.EBX[1:0] before perfmon v6
The UMASK2 and EQ bits in IA32_PERFEVTSELx MSRs are available in PMU
v6 and up.  If the guest PMU version is configured to be less than 6,
these two bits need to be removed from config_mask, so that KVM won't
expose them to the guest.

Changed the condition to clear kvm_pmu_cap.arch_pebs from
!enable_mediated_pmu to '(kvm_pmu_cap.version < 6)' which correctly
indicate that it doesn't exist prior to version 6.

Fixes: a89023564cc9 ("KVM: x86/pmu: Add archPerfmonExt leaf emulation")
Signed-off-by: Zide Chen <zide.chen@intel.com>