crypto: qat - enable qat_6xxx driver for power on
The GEN6 device with revision ID A0 does not support the wireless
algorithm. To prevent the current driver usage with revision ID B0,
the check was added to return an error if the SOC does not support
wireless.

Updated the logic to skip the check if it is an AO SOC as non-wireless
support is present.

This commit should not be pushed to upstream.

Signed-off-by: Suman Kumar Chakraborty <suman.kumar.chakraborty@intel.com>