intel_vsec: Add S3M driver dependencies to VSEC driver
The VSEC driver requires the S3M ID value to be specified in order to
probe the S3M aux device. An indexing bit was also specified to denote
S3M capabilities are present.

If the VSEC driver detects the presence of a DVSEC register containing
S3M functionality and the S3M driver is registered, s3m_probe will be
called.

Signed-off-by: Will Skrydlak <will.j.skrydlak@linux.intel.com>