crypto: qat - enable RAS support for GEN6 devices
RAS (Reliability, Accessibility, Serviceability) enables the qat_6xxx
driver to take appropriate actions depending upon the severity of the
error.
The actions range from a simple notification to the user to a function
level reset. The errors are read in error source registers referenced
ERRSOU0 to ERRSOU3. Each occurrence of an error increments a counter
specific to the error type.

Add the infrastructure to handle the errors from ERRSOU0 to ERRSOU3.
Enable the RAS error counters in sysfs for GEN6 HW

The counters added are:
- errors_correctable - Hardware correctable errors include the
  error conditions where the system can recover without any loss of
  information.
- errors_nonfatal - This category of errors includes the errors that
  can be isolated to one or more in-flight requests.
- errors_fatal - This category of errors includes the errors that
  cannot be contained. An FLR is required when such errors occur.

Signed-off-by: Suman Kumar Chakraborty <suman.kumar.chakraborty@intel.com>