crypto: qat - add event handling for uacce transport layer
This patch adds event handling to qat's uacce interface. The events
are exposed to the user space via sysfs attribute 'dev_event'.
The attribute leverages the sysfs_notify() function to allow the userspace
to poll for events.

Supported events are:
- restarting (device is being restarted)
- restarted (device restart completed)
- error (detected error)

User space must release acquired resources in response to the
restarting event.

Co-developed-by: Michal Witwicki <michal.witwicki@intel.com>
Signed-off-by: Michal Witwicki <michal.witwicki@intel.com>

Signed-off-by: Damian Muszynski <damian.muszynski@intel.com>