tboot: Disable FRED before tboot's shutdown callback
Disable FRED along with the other bits we're clearing in CR4.  This
fixes a crash in the tboot's shutdown entry code when FRED is used
in linux.  This makes sense because the first thing that code does is
switch to compatibility mode and then clear CR0.PG and EFER.LME:

https://github.com/mateusz-mowka/tboot-lcptools-v2-with-lms/blob/main/tboot/common/shutdown.S#L135

...all of which are illegal with FRED enabled, for example FRED
spec revision 4.0 section 7.2 says:

"In addition, enabling FRED transitions disables the ability of software
to enter compatibility mode in ring 0. An execution of far CALL, IRET,
far JMP, or far RET causes a transition to compatibility mode if it
loads CS from a code-segment descriptor in which the L bit is 0. (...)
When FRED transitions are enabled, such an execution in ring 0 causes #GP."

Signed-off-by: Andrew Zaborowski <andrew.zaborowski@intel.com>