EDAC/imh: Fix decoded memory controller index zero issue
The current imh_edac driver for Intel Diamond Rapids always reports
memory errors as originating from the first memory controller, even
when they are from other controllers. This is due to a missed
initialization of the mapping from the physical index to the logical
index of memory controllers.

Fix this issue by initializing the memory controller index mapping table.

JIRA to track this bug:

  https://jira.devtools.intel.com/browse/LFE-17339

Signed-off-by: Qiuxu Zhuo <qiuxu.zhuo@intel.com>