x86/mm/erdt: Fix the clang warning
In the default branch of the switch statement,
there are no statements before the closing brace.
This syntax is only allowed in the C23 standard,
while the current compiler may be using an earlier
C standard.
Add a break; statement in the default branch.

Note: the ERDT code has been re-implemented in
the latest intel-next.

Signed-off-by: Chen Yu <yu.c.chen@intel.com>