crypto: qat - distribute DCC service in SVM mode
This patch refactors the QAT SVM ring pair distribution logic to support
the DCC (Data Compression and Cryptography) service. Service type fields
are updated to u32 to handle enums for both base and extended services.

Since the DCC service cannot be combined with other services, this
property is used to identify the type of service available in the bank.

Signed-off-by: Michal Witwicki <michal.witwicki@intel.com>